+++ CHANGELOG.md
@@ -14,6 +14,8 @@ x.x.x Release notes (yyyy-MM-dd)
 
 * The bowser will no longer show objects that have no persisted properties.
 * `RLMSchema`, `RLMObjectSchema`, & `RLMProperty` now have more useful descriptions.
+* Opening an encrypted Realm while a debugger is attached to the process no
+  longer throws an exception.
 
 ### Bugfixes